AG-13306 Create NavigatorDOMProxy class
