+++ packages/ag-charts-community/src/chart/navigator/navigator.ts
@@ -3,11 +3,8 @@ import { BaseModuleInstance } from '../../module/module';
 import type { ModuleContext } from '../../module/moduleContext';
 import type { BBox } from '../../scene/bbox';
 import type { Group } from '../../scene/group';
-import { setAttribute } from '../../util/attributeUtil';
 import type { BBoxProvider } from '../../util/bboxinterface';
 import { setElementBBox } from '../../util/dom';
-import { formatPercent } from '../../util/format.util';
-import { initToolbarKeyNav } from '../../util/keynavUtil';
 import { clamp } from '../../util/number';
 import { ObserveChanges } from '../../util/proxy';
 import { BOOLEAN, OBJECT, POSITIVE_NUMBER, Validate } from '../../util/validation';
@@ -15,6 +12,7 @@ import { InteractionState, type PointerInteractionEvent } from '../interaction/i
 import type { RegionEvent } from '../interaction/regionManager';
 import type { ZoomChangeEvent } from '../interaction/zoomManager';
 import { type LayoutCompleteEvent, LayoutElement } from '../layout/layoutManager';
+import { NavigatorDOMProxy } from './navigatorDOMProxy';
 import { RangeHandle } from './shapes/rangeHandle';
 import { RangeMask } from './shapes/rangeMask';
 import { RangeSelector } from './shapes/rangeSelector';
@@ -56,13 +54,7 @@ export class Navigator extends BaseModuleInstance implements ModuleInstance {
 
     private dragging?: NavigatorButtonType;
     private panStart?: number;
-    private _min = 0;
-    private _max = 1;
-
-    private readonly minRange = 0.001;
-
-    private readonly proxyNavigatorToolbar: HTMLElement;
-    private readonly proxyNavigatorElements: [HTMLInputElement, HTMLInputElement, HTMLInputElement];
+    private readonly domProxy: NavigatorDOMProxy;
 
     constructor(private readonly ctx: ModuleContext) {
         super();
@@ -82,54 +74,8 @@ export class Navigator extends BaseModuleInstance implements ModuleInstance {
             ctx.zoomManager.addListener('zoom-change', (event) => this.onZoomChange(event))
         );
 
-        this.proxyNavigatorToolbar = this.ctx.proxyInteractionService.createProxyContainer({
-            type: 'toolbar',
-            id: `navigator-toolbar`,
-            classList: ['ag-charts-proxy-navigator-toolbar'],
-            ariaOrientation: 'vertical',
-            ariaLabel: { id: 'ariaLabelNavigator' },
-        });
+        this.domProxy = new NavigatorDOMProxy(ctx);
         this.updateGroupVisibility();
-
-        this.proxyNavigatorElements = [
-            this.ctx.proxyInteractionService.createProxyElement({
-                type: 'slider',
-                id: 'ag-charts-navigator-min',
-                ariaLabel: { id: 'ariaLabelNavigatorMinimum' },
-                ariaOrientation: 'horizontal',
-                parent: this.proxyNavigatorToolbar,
-                onchange: (ev) => this.onMinSliderChange(ev),
-            }),
-            this.ctx.proxyInteractionService.createProxyElement({
-                type: 'slider',
-                id: 'ag-charts-navigator-pan',
-                ariaLabel: { id: 'ariaLabelNavigatorRange' },
-                ariaOrientation: 'horizontal',
-                parent: this.proxyNavigatorToolbar,
-                onchange: (ev) => this.onPanSliderChange(ev),
-            }),
-            this.ctx.proxyInteractionService.createProxyElement({
-                type: 'slider',
-                id: 'ag-charts-navigator-max',
-                ariaLabel: { id: 'ariaLabelNavigatorMaximum' },
-                ariaOrientation: 'horizontal',
-                parent: this.proxyNavigatorToolbar,
-                onchange: (ev) => this.onMaxSliderChange(ev),
-            }),
-        ];
-        this.proxyNavigatorElements.forEach((slider) => setAttribute(slider, 'data-preventdefault', false));
-        this.setSliderRatio(this.proxyNavigatorElements[0], this._min);
-        this.setSliderRatio(this.proxyNavigatorElements[2], this._max);
-        this.setPanSliderValue(this._min, this._max);
-        initToolbarKeyNav({
-            orientation: 'vertical',
-            toolbar: this.proxyNavigatorToolbar,
-            buttons: this.proxyNavigatorElements,
-        });
-        this.destroyFns.push(() => {
-            this.proxyNavigatorElements.forEach((e) => e.remove());
-            this.proxyNavigatorToolbar.remove();
-        });
     }
 
     public updateBackground(oldGroup?: Group, newGroup?: Group) {
@@ -141,7 +87,7 @@ export class Navigator extends BaseModuleInstance implements ModuleInstance {
 
         if (this.rangeSelector == null || enabled === this.rangeSelector.visible) return;
         this.rangeSelector.visible = enabled;
-        this.proxyNavigatorToolbar.ariaHidden = (!enabled).toString();
+        this.domProxy.proxyNavigatorToolbar.ariaHidden = (!enabled).toString();
 
         if (enabled) {
             this.updateZoom();
@@ -167,10 +113,10 @@ export class Navigator extends BaseModuleInstance implements ModuleInstance {
         if (this.enabled) {
             const { y, height } = this;
             this.layoutNodes(x, y, width, height);
-            setElementBBox(this.proxyNavigatorToolbar, { x, y, width, height });
-            this.proxyNavigatorToolbar.style.removeProperty('display');
+            setElementBBox(this.domProxy.proxyNavigatorToolbar, { x, y, width, height });
+            this.domProxy.proxyNavigatorToolbar.style.removeProperty('display');
         } else {
-            this.proxyNavigatorToolbar.style.display = 'none';
+            this.domProxy.proxyNavigatorToolbar.style.display = 'none';
         }
 
         this.x = x;
@@ -204,7 +150,8 @@ export class Navigator extends BaseModuleInstance implements ModuleInstance {
         if (!this.enabled) return;
         this.updateCursor(event);
 
-        const { mask, minHandle, maxHandle, x, width, _min: min } = this;
+        const { mask, minHandle, maxHandle, x, width } = this;
+        const { _min: min } = this.domProxy;
         const { regionOffsetX, regionOffsetY } = event;
 
         if (minHandle.zIndex < maxHandle.zIndex) {
@@ -232,8 +179,8 @@ export class Navigator extends BaseModuleInstance implements ModuleInstance {
     private onDrag(event: RegionEvent<'drag'>) {
         if (!this.enabled || this.dragging == null) return;
 
-        const { dragging, minRange, panStart, x, width } = this;
-        let { _min: min, _max: max } = this;
+        const { dragging, panStart, x, width } = this;
+        let { minRange, _min: min, _max: max } = this.domProxy;
         const { regionOffsetX } = event;
 
         const ratio = (regionOffsetX - x) / width;
@@ -248,8 +195,8 @@ export class Navigator extends BaseModuleInstance implements ModuleInstance {
             max = min + span;
         }
 
-        this._min = min;
-        this._max = max;
+        this.domProxy._min = min;
+        this.domProxy._max = max;
 
         this.updateZoom();
     }
@@ -267,60 +214,17 @@ export class Navigator extends BaseModuleInstance implements ModuleInstance {
         const { x } = event;
         if (!x) return;
 
-        this._min = x.min;
-        this._max = x.max;
+        this.domProxy._min = x.min;
+        this.domProxy._max = x.max;
         this.updateNodes(x.min, x.max);
-        this.setPanSliderValue(x.min, x.max);
-        this.setSliderRatio(this.proxyNavigatorElements[0], x.min);
-        this.setSliderRatio(this.proxyNavigatorElements[2], x.max);
-    }
-
-    private onPanSliderChange(_event: Event) {
-        const ratio = this.getSliderRatio(this.proxyNavigatorElements[1]);
-        const span = this._max - this._min;
-        this._min = clamp(0, ratio, 1 - span);
-        this._max = this._min + span;
-        this.updateZoom();
-    }
-
-    private onMinSliderChange(_event: Event) {
-        const slider = this.proxyNavigatorElements[0];
-        this._min = this.setSliderRatioClamped(slider, 0, this._max - this.minRange);
-        this.updateZoom();
-    }
-
-    private onMaxSliderChange(_event: Event) {
-        const slider = this.proxyNavigatorElements[2];
-        this._max = this.setSliderRatioClamped(slider, this._min + this.minRange, 1);
-        this.updateZoom();
-    }
-
-    private setPanSliderValue(min: number, max: number) {
-        this.proxyNavigatorElements[1].value = `${Math.round(min * 100)}`;
-        this.proxyNavigatorElements[1].ariaValueText = this.ctx.localeManager.t('ariaValuePanRange', { min, max });
-    }
-
-    private setSliderRatioClamped(slider: HTMLInputElement, clampMin: number, clampMax: number) {
-        const ratio = this.getSliderRatio(slider);
-        const clampedRatio = clamp(clampMin, ratio, clampMax);
-        if (clampedRatio !== ratio) {
-            this.setSliderRatio(slider, clampedRatio);
-        }
-        return clampedRatio;
-    }
-
-    private setSliderRatio(slider: HTMLInputElement, ratio: number) {
-        const value = Math.round(ratio * 100);
-        slider.value = `${value}`;
-        slider.ariaValueText = formatPercent(value / 100);
-    }
-
-    private getSliderRatio(slider: HTMLInputElement) {
-        return parseFloat(slider.value) / 100;
+        this.domProxy.setPanSliderValue(x.min, x.max);
+        this.domProxy.setSliderRatio(this.domProxy.proxyNavigatorElements[0], x.min);
+        this.domProxy.setSliderRatio(this.domProxy.proxyNavigatorElements[2], x.max);
     }
 
     private layoutNodes(x: number, y: number, width: number, height: number) {
-        const { rangeSelector, mask, minHandle, maxHandle, _min: min, _max: max } = this;
+        const { rangeSelector, mask, minHandle, maxHandle } = this;
+        const { _min: min, _max: max } = this.domProxy;
 
         rangeSelector.layout(x, y, width, height, minHandle.width / 2, maxHandle.width / 2);
         mask.layout(x, y, width, height);
@@ -338,7 +242,7 @@ export class Navigator extends BaseModuleInstance implements ModuleInstance {
         [minHandle, this.maskVisibleRange, maxHandle].forEach((node, index) => {
             const bbox = node.getBBox();
             const tbox = { x: bbox.x - x, y: bbox.y - y, height: bbox.height, width: bbox.width };
-            setElementBBox(this.proxyNavigatorElements[index], tbox);
+            setElementBBox(this.domProxy.proxyNavigatorElements[index], tbox);
         });
     }
 
@@ -348,10 +252,6 @@ export class Navigator extends BaseModuleInstance implements ModuleInstance {
 
     private updateZoom() {
         if (!this.enabled) return;
-
-        const { _min: min, _max: max } = this;
-        if (min == null || max == null) return;
-
-        return this.ctx.zoomManager.updateZoom('navigator', { x: { min, max } });
+        this.domProxy.updateZoom();
     }
 }
